Database completed Run Service Completed HIIT Service needs notification ViewModel completed for settings only UI started

+++ app/src/main/java/com/example/workout_journal/data/entity/WeightSet.kt
@@ -18,10 +18,10 @@ import androidx.room.PrimaryKey
     indices = [Index("weightExerciseId")]
 )
 data class WeightSet(
-    @PrimaryKey(autoGenerate = true) val id : Int =0,
-    val weightExerciseId: Int,
+    @PrimaryKey(autoGenerate = true) val id : Long =0,
+    val weightExerciseId: Long,
     val set : Int,
     val reps : Int,
-    val weightKg : Float,
+    val weightKg : Double,
     val setType: SetType
     )
